pyproject.toml
[project]
name = "novelback"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "configdict>=2.13.4",
    "fastapi>=0.128.1",
    "fields>=5.0.0",
    "openai>=2.17.0",
    "optional>=0.0.1",
    "python-multipart>=0.0.22",
    "tortoise-orm>=0.25.4",
    "uvicorn[standard]>=0.40.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=models --cov=controllers --cov=api --cov-report=term-missing"
testpaths = [
    "test",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "real_llm: 需要真实 LLM API 的测试（需配置 test/.test.env）",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pyinstrument>=5.1.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
